+++ backend/src/handlers/actions/PokerTableJoinHandler.ts
@@ -3,7 +3,7 @@ import type { Response } from 'express';
 import type { PokerTableJoinPayload, PokerTableJoinOutput } from '../../shared/api/types/PokerTableJoin';
 
 // Internal
-import { BaseHandler } from '../../shared/BaseHandler';
+import { BaseHandler } from '../BaseHandler';
 import { Result } from '@shared/Result';
 import { Rooms } from '../../sockets/Rooms';
 import { GameLobbyService } from '../../game-lobby-service';
